refactor(texts): name folder
