+++ src/modules/texts/TextEdit/components/Markdown/services/fetchPatchTextService.ts
@@ -1,6 +1,6 @@
 import { IFetchResponseDefault } from "shared/common/types/Fetch";
 import api from "shared/services/api";
-import { IFetchPatchTextService } from "modules/texts/TextEdit/components/Markdown/service/types";
+import { IFetchPatchTextService } from "modules/texts/TextEdit/components/Markdown/services/types";
 
 export const fetchPatchTextService: IFetchPatchTextService = async ({
 	params,
